Update pdffont.js
To fix the problem, we need to ensure that all occurrences of the specified substrings are replaced in the flashEncode function. This can be achieved by using regular expressions with the global flag (g) instead of string literals in the replace method. This change will ensure that all occurrences of the substrings are replaced, not just the first one.